a clumsy separation of an overloaded object
so SchemaGuesser was responsible for both guessing schemas and
connecting to Mongo. the whole reason I'm extracting this from a
previous project is I have some similar stuff I want to do, but with
some data which isn't in Mongo.

so I tore out the Mongo-connecting bits. this thing is still a mess,
to say the least, but it makes a little more sense, and more to the
point, I'll be able to actually use it to analyze arbitrary JSON, not
just arbitrary JSON from Mongo.